+++ royal/testing.py
@@ -1,22 +1,15 @@
 import functools
 
-from pyramid.testing import DummyResource
 
-
-class _DummyResourceBase(DummyResource):
+class _DummyResourceBase(object):
 
     def __init__(self, name, parent, request, resource_path, resource_type,
                  **kw):
-        super(_DummyResourceBase, self).__init__(
-            name,
-            parent,
-            request=request,
-            **kw
-            )
-        self.name = name
-        self.parent = parent
+        self.__name__ = self.name = name
+        self.__parent__ = parent
         self.__resource_path__ = resource_path
         self.__resource_type__ = resource_type
+        self.__dict__.update(kw)
 
 
 DummyCollection = functools.partial(_DummyResourceBase,